requirements.txt
PyJWT==2.9.0
pandas==2.2.3
Flask==3.0.3
cryptography==43.0.3